Update employe model, view, and controller
This commit updates the `Employe` model by adding the `activo` attribute. It also updates the `create-employe.blade.php` view by fixing a typo in the "Puesto" label. Additionally, it updates the `EmployeController.php` controller by adding validation logic to the `update` method. Finally, it updates the `fecha_contratacion` attribute in the `casts` array to use the `date` format.